+++ backend/backend/serializers.py
@@ -1,0 +1,11 @@
+from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
+from rest_framework.exceptions import AuthenticationFailed
+from django.utils.translation import gettext_lazy as _
+
+# Validar token
+class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
+    def validate(self, attrs):
+        try:
+            return super().validate(attrs)
+        except AuthenticationFailed as e:
+            raise AuthenticationFailed(_("Las credenciales son incorrectas."))